Support Query Options with Stream
There are several options which could not be passed in that are
handy to have for debugging as well as tuning.  That list is
currently:

 * timeout
 * pool_timeout
 * log

This documentation can be found here:
https://hexdocs.pm/ecto/Ecto.Repo.html#module-shared-options

This allows users of the library to pass in these as options to
the stream as well as the other supported options the stream itself
supports.